backend/core/security.py
"""
Модуль безопасности.
Работа с паролями, JWT-токенами и авторизацией.
Используем bcrypt напрямую для совместимости с Python 3.14.
"""

from datetime import datetime, timedelta
from typing import Optional

import bcrypt
from jose import JWTError, jwt

from .config import get_settings


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Проверяем, совпадает ли пароль с хешем.
    
    Args:
        plain_password: Пароль в открытом виде
        hashed_password: Хеш пароля из базы
        
    Returns:
        True если пароль верный, иначе False
    """
    password_bytes = plain_password.encode('utf-8')
    hash_bytes = hashed_password.encode('utf-8')
    return bcrypt.checkpw(password_bytes, hash_bytes)


def get_password_hash(password: str) -> str:
    """
    Хешируем пароль для безопасного хранения в БД.
    
    Args:
        password: Пароль в открытом виде
        
    Returns:
        Хеш пароля
    """
    password_bytes = password.encode('utf-8')
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password_bytes, salt)
    return hashed.decode('utf-8')


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    Создаём JWT-токен для авторизации.
    
    Args:
        data: Данные для включения в токен (обычно user_id или email)
        expires_delta: Время жизни токена (опционально)
        
    Returns:
        Закодированный JWT-токен
    """
    settings = get_settings()
    to_encode = data.copy()
    
    # Устанавливаем время истечения
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.access_token_expire_minutes)
    
    to_encode.update({"exp": expire})
    
    # Кодируем и подписываем токен
    encoded_jwt = jwt.encode(
        to_encode, 
        settings.secret_key, 
        algorithm=settings.algorithm
    )
    return encoded_jwt


def decode_token(token: str) -> Optional[dict]:
    """
    Декодируем и проверяем JWT-токен.
    
    Args:
        token: JWT-токен из заголовка Authorization
        
    Returns:
        Расшифрованные данные или None если токен невалидный
    """
    settings = get_settings()
    try:
        payload = jwt.decode(
            token, 
            settings.secret_key, 
            algorithms=[settings.algorithm]
        )
        return payload
    except JWTError:
        # Токен невалидный или истёк
        return None
