backend/requirements.txt
# Основные зависимости для ERP-lite системы
# Сервер и фреймворк
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ORM и работа с базой данных
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Валидация данных и настройки
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Безопасность: JWT-токены и хеширование паролей
python-jose[cryptography]==3.3.0
bcrypt>=5.0.0

# Обработка форм и файлов
python-multipart==0.0.6

# Работа с датами
python-dateutil==2.8.2
